requirements.txt
python-telegram-bot
validators
yt-dlp